complete coupon functionality

+++ app/controllers/orders_controller.rb
@@ -34,6 +34,7 @@ class OrdersController <ApplicationController
 
   def show
     @order = Order.find(params[:order_id])
+    @coupon = Coupon.find_by(name: @order.coupon_code)
     @address = @order.address
     user = User.find(session[:user_id])
     @addresses = user.addresses
@@ -54,9 +55,26 @@ class OrdersController <ApplicationController
     address = Address.find(params[:address].to_i)
     order = user.orders.create(address: address)
     create_item_orders(order)
-    session.delete(:cart)
-    redirect_to "/profile/orders"
-    flash[:success] = "Thank You For Your Order!"
+    order.update(grand_total: order.grandtotal)
+    if coupon = Coupon.find_by(name: params[:coupon_code])
+      if coupon.has_order_items?(order)
+        order.update(coupon_code: params[:coupon_code])
+        order.update_coupon_discounts
+        session.delete(:cart)
+        redirect_to "/profile/orders"
+        flash[:success] = "Your coupon was applied to your grand total homie"
+      else
+        flash[:error] = "Not a valid code, yo. Try again."
+        redirect_to "/orders/new"
+      end
+    elsif !params[:coupon_code].nil?
+      session.delete(:cart)
+      redirect_to "/profile/orders"
+      flash[:success] = "Thank You For Your Order!"
+    else
+      flash[:error] = "Not a valid code, yo. Try again."
+      redirect_to "/orders/new"
+    end
   end
 
   def ship
